linkarr-backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Caching & Queue
redis==5.0.1
celery==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.1
aiohttp==3.9.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2

# External APIs
requests==2.31.0
